% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveFunctions.R
\name{batPicClassifier}
\alias{batPicClassifier}
\title{batPicClassifier}
\usage{
batPicClassifier(option.list = list(shiny.launch.browser = TRUE,
  shiny.maxRequestSize = 100 * 1024^2, encoding = "UTF-8"))
}
\arguments{
\item{option.list}{a list of global shiny options}
}
\value{
launches a WebApp
}
\description{
launches an interactive web app
}
\seealso{
Other res: \code{\link{changeSelectize}}
}

