% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{changeSelectize}
\alias{changeSelectize}
\title{changeSelectize}
\usage{
changeSelectize(dir, values, selected)
}
\arguments{
\item{dir}{+1 or -1 for next or previous element}

\item{values}{a vector of values that can be selectegd}

\item{selected}{the currently selected value}
}
\value{
a value
}
\description{
helper function to change Selectize input by 1
}
\seealso{
Other res: \code{\link{batPicClassifier}}
}

